Capstone Poised Project
This project is a Java program.
This project is a management system for a small structural engineering firm called 'Poised'. The main function of the system is to keep track of the many projects on which they work and also capture about new projects.
The program stores information like:
● Project number.
● Project name.
● What type of building is being designed? E.g. House, apartment block or
store, etc.
● The physical address for the project.
● ERF number.
● The total fee being charged for the project.
● The total amount paid to date.
● Deadline for the project.
● The name, telephone number, email address and physical address of the
architect for the project.
● The name, telephone number, email address and physical address of the
contractor for the project.
● The name, telephone number, email address and physical address of the
customer for the project.