Add files via upload
This file is a k-mean algorithm about the birth rate and life expectancy measured for each country given in the case study.
As you continue assessing the code, the second phase of the program, the squared distance is calculated between each point and the centroid to which the point belonged. The sum of all the squared distances is also calculated.